# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AppConfigDTO:
    properties:
      directives:
        description: 指令配置
        type: string
      logFile:
        description: 日志文件
        type: string
      logFormat:
        description: 日志格式
        type: string
      logLevel:
        description: 日志级别
        type: string
      name:
        description: 应用名称
        type: string
      transactionTTL:
        description: 事务超时时间(毫秒)
        example: 60000
        type: integer
    type: object
  dto.AppConfigPatchDTO:
    properties:
      directives:
        description: 指令配置
        type: string
      logFile:
        description: 日志文件
        example: /dev/stdout
        type: string
      logFormat:
        description: 日志格式
        example: console
        type: string
      logLevel:
        description: 日志级别
        example: info
        type: string
      name:
        description: 应用名称
        example: coraza
        type: string
      transactionTTL:
        description: 事务超时时间(毫秒)
        example: 60000
        type: integer
    type: object
...
  dto.ConfigPatchRequest:
    description: 用于部分更新配置的请求参数
    properties:
      engine:
        allOf:
        - $ref: '#/definitions/dto.EnginePatchDTO'
        description: 引擎配置
      haproxy:
        allOf:
        - $ref: '#/definitions/dto.HaproxyPatchDTO'
        description: HAProxy配置
      isDebug:
        description: 是否开启调试模式
        example: false
        type: boolean
      isResponseCheck:
        description: 是否检查响应
        example: false
        type: boolean
      name:
        description: 配置名称
        example: AppConfig
        type: string
    type: object
  dto.ConfigResponse:
    description: 配置响应
    properties:
      createdAt:
        description: 创建时间
        type: string
      engine:
        allOf:
        - $ref: '#/definitions/dto.EngineDTO'
        description: 引擎配置
      haproxy:
        allOf:
        - $ref: '#/definitions/dto.HaproxyDTO'
        description: HAProxy配置
      id:
        description: 配置ID
        type: string
      isDebug:
        description: 是否开启调试模式
        type: boolean
      isResponseCheck:
        description: 是否检查响应
        type: boolean
      name:
        description: 配置名称
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  dto.EngineDTO:
    properties:
      appConfig:
        description: 应用配置列表
        items:
          $ref: '#/definitions/dto.AppConfigDTO'
        type: array
      bind:
        description: 引擎绑定地址
        type: string
      useBuiltinRules:
        description: 是否使用内置规则
        type: boolean
    type: object
  dto.EnginePatchDTO:
    properties:
      appConfig:
        description: 应用配置列表
        items:
          $ref: '#/definitions/dto.AppConfigPatchDTO'
        type: array
      bind:
        description: 引擎绑定地址
        example: 127.0.0.1:2342
        type: string
      useBuiltinRules:
        description: 是否使用内置规则
        example: true
        type: boolean
    type: object
  dto.GetUserInfoResponseData:
    properties:
      id:
        example: "1234567890"
        type: string
      needReset:
        example: false
        type: boolean
      role:
        example: admin
        type: string
      username:
        example: user123
        type: string
    type: object
  dto.HaproxyDTO:
    properties:
      backupsNumber:
        description: 备份数量
        type: integer
      configBaseDir:
        description: 配置文件根目录
        type: string
      haproxyBin:
        description: HAProxy二进制文件路径
        type: string
      spoeAgentAddr:
        description: SPOE代理地址
        type: string
      spoeAgentPort:
        description: SPOE代理端口
        type: integer
      thread:
        description: 线程数
        type: integer
    type: object
  dto.HaproxyPatchDTO:
    properties:
      backupsNumber:
        description: 备份数量
        example: 5
        type: integer
      configBaseDir:
        description: 配置文件根目录
        example: /simple-waf
        type: string
      haproxyBin:
        description: HAProxy二进制文件路径
        example: haproxy
        type: string
      spoeAgentAddr:
        description: SPOE代理地址
        example: 127.0.0.1
        type: string
      spoeAgentPort:
        description: SPOE代理端口
        example: 2342
        type: integer
      thread:
        description: 线程数
        example: 4
        type: integer
    type: object
  dto.LoginResponseData:
    properties:
      token:
        description: JWT token
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 用户信息
    type: object
  dto.ResetPasswordResponseData:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.User'
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  dto.UserCreateRequest:
    properties:
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - user
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - password
    - role
    - username
    type: object
  dto.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UserPasswordResetRequest:
    properties:
      newPassword:
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  dto.UserUpdateRequest:
    properties:
      needReset:
        type: boolean
      password:
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - auditor
        - configurator
        - user
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    type: object
  model.APIResponse:
    description: API响应的标准格式
    properties:
      code:
        description: HTTP状态码
        example: 200
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息(当success=false时)
        example: 参数错误
        type: string
      message:
        description: 响应消息
        example: 操作成功
        type: string
      requestId:
        description: 请求ID，用于跟踪
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        description: 是否成功
        example: true
        type: boolean
      timestamp:
        description: 响应时间戳
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.ErrResponse:
    description: 错误的API响应标准格式
    properties:
      code:
        example: 400
        type: integer
      error:
        example: 参数错误
        type: string
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.ErrResponseDontShowError:
    description: 错误的API响应标准格式,不展示 error
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.SuccessResponseNoData:
    description: 成功的API响应标准格式,没有 data
    properties:
      code:
        example: 200
        type: integer
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastLogin:
        type: string
      needReset:
        description: 是否需要重置密码
        type: boolean
      permissions:
        description: 额外权限
        items:
          type: string
        type: array
      role:
        description: 用户角色
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 简单的 Web 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple-WAF API
  version: "1.0"
paths:
...
  /api/v1/config:
    get:
      description: 获取当前系统配置信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取配置成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ConfigResponse'
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取系统配置
      tags:
      - 配置管理
    patch:
      consumes:
      - application/json
      description: 使用补丁方式更新系统配置
      parameters:
      - description: 配置更新信息
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/dto.ConfigPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 配置更新成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ConfigResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 更新系统配置
      tags:
      - 配置管理
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录并获取JWT令牌
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponseData'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      summary: 用户登录
      tags:
      - 认证
  /auth/me:
    get:
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户信息成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetUserInfoResponseData'
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: 用户重置自己的密码
      parameters:
      - description: 密码重置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "400":
          description: 请求参数错误或原密码错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 重置密码
      tags:
      - 认证
  /users:
    get:
      description: 获取系统中所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户列表成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取所有用户
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 管理员创建新用户
      parameters:
      - description: 用户创建信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户创建成功
          schema:
            $ref: '#/definitions/dto.ResetPasswordResponseData'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "409":
          description: 用户名已存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 创建新用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      description: 管理员删除指定用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户删除成功
          schema:
            $ref: '#/definitions/model.APIResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 管理员更新指定用户的信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户更新成功
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.APIResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.APIResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"